Refactor

+++ src/functions/activation.jl
@@ -71,7 +71,7 @@ function ∇elu!(y::Array{T}, gy::Array{T}, x::Array{T}, gx::Array{T}) where T
 end
 
 doc"""
-    leaky_relu(x::Var, alpha::Float64=0.1)
+    leaky_relu(x::Var, alpha=0.1)
 
 Leaky Rectified Linear Unit.
 
@@ -86,7 +86,7 @@ x & x > 0 \\
 # References
 * Maas et al., ["Rectifier Nonlinearities Improve Neural Network Acoustic Models"](http://web.stanford.edu/~awni/papers/relu_hybrid_icml2013_final.pdf), ICML 2013.
 """
-leaky_relu(x::Var, alpha::Float64=0.1) = Var(leaky_relu.(x.data,eltype(x)(alpha)), (leaky_relu,x,alpha))
+leaky_relu(x::Var, alpha=0.1) = Var(leaky_relu.(x.data,eltype(x)(alpha)), (leaky_relu,x,alpha))
 leaky_relu(x::Node; name="") = Node(leaky_relu, (x,), name)
 leaky_relu(x::T, alpha::T) where T = x >= T(0) ? x : x*alpha
